[New] Created test cases for pages-admin-form-builder-custom-field-dialog component from enzyme to rtl
